fixup! fixup! fixup! [move] Switch to public(script) for Sui Move entry points
