+++ sui_programmability/adapter/src/adapter.rs
@@ -28,7 +28,7 @@ use sui_verifier::{
 
 use move_core_types::{
     account_address::AccountAddress,
-    identifier::{IdentStr, Identifier},
+    identifier::Identifier,
     language_storage::{ModuleId, StructTag, TypeTag},
     resolver::{ModuleResolver, ResourceResolver},
     value::MoveTypeLayout,
@@ -804,7 +804,7 @@ fn type_check_struct(
         Err(SuiError::TypeError {
             error: format!(
                 "Expected argument of type {}, but found type {}",
-                format_signature_token(module, param_type),
+                sui_verifier::format_signature_token(module, param_type),
                 arg_type
             ),
         })
@@ -872,7 +872,7 @@ fn struct_tag_equals_struct_inst(
     param_type: StructHandleIndex,
     param_type_arguments: &[SignatureToken],
 ) -> bool {
-    let (address, module_name, struct_name) = resolve_struct(module, param_type);
+    let (address, module_name, struct_name) = sui_verifier::resolve_struct(module, param_type);
 
     // same address
     &arg_type.address == address
@@ -892,66 +892,4 @@ fn struct_tag_equals_struct_inst(
                 )
             },
         )
-}
-
-fn resolve_struct(
-    module: &CompiledModule,
-    sidx: StructHandleIndex,
-) -> (&AccountAddress, &IdentStr, &IdentStr) {
-    let shandle = module.struct_handle_at(sidx);
-    let mhandle = module.module_handle_at(shandle.module);
-    let address = module.address_identifier_at(mhandle.address);
-    let module_name = module.identifier_at(mhandle.name);
-    let struct_name = module.identifier_at(shandle.name);
-    (address, module_name, struct_name)
-}
-
-fn format_signature_token(module: &CompiledModule, t: &SignatureToken) -> String {
-    match t {
-        SignatureToken::Bool => "bool".to_string(),
-        SignatureToken::U8 => "u8".to_string(),
-        SignatureToken::U64 => "u64".to_string(),
-        SignatureToken::U128 => "u128".to_string(),
-        SignatureToken::Address => "address".to_string(),
-        SignatureToken::Signer => "signer".to_string(),
-        SignatureToken::Vector(inner) => {
-            format!("vector<{}>", format_signature_token(module, inner))
-        }
-        SignatureToken::Reference(inner) => format!("&{}", format_signature_token(module, inner)),
-        SignatureToken::MutableReference(inner) => {
-            format!("&mut {}", format_signature_token(module, inner))
-        }
-        SignatureToken::TypeParameter(i) => format!("T{}", i),
-
-        SignatureToken::Struct(idx) => format_signature_token_struct(module, *idx, &[]),
-        SignatureToken::StructInstantiation(idx, ty_args) => {
-            format_signature_token_struct(module, *idx, ty_args)
-        }
-    }
-}
-
-fn format_signature_token_struct(
-    module: &CompiledModule,
-    sidx: StructHandleIndex,
-    ty_args: &[SignatureToken],
-) -> String {
-    let (address, module_name, struct_name) = resolve_struct(module, sidx);
-    let s;
-    let ty_args_string = if ty_args.is_empty() {
-        ""
-    } else {
-        s = ty_args
-            .iter()
-            .map(|t| format_signature_token(module, t))
-            .collect::<Vec<_>>()
-            .join(", ");
-        &s
-    };
-    format!(
-        "0x{}::{}::{}{}",
-        address.short_str_lossless(),
-        module_name,
-        struct_name,
-        ty_args_string
-    )
 }
