+++ sui_programmability/adapter/src/unit_tests/data/publish_init_public/sources/M1.move
@@ -12,7 +12,7 @@ module Test::M1 {
     }
 
     // public initializer - should not be executed
-    public fun init(ctx: &mut TxContext) {
+    public(script) fun init(ctx: &mut TxContext) {
         let value = 42;
         let singleton = Object { id: TxContext::new_id(ctx), value };
         Transfer::transfer(singleton, TxContext::sender(ctx))
