+++ sui_programmability/examples/basics/sources/Lock.move
@@ -41,7 +41,7 @@ module Basics::Lock {
 
     /// Lock some content inside a shared object. A Key is created and is
     /// sent to the transaction sender.
-    public fun create<T: store + key>(obj: T, ctx: &mut TxContext) {
+    public(script) fun create<T: store + key>(obj: T, ctx: &mut TxContext) {
         let id = TxContext::new_id(ctx);
         let for = *ID::inner(&id);
 
@@ -58,7 +58,7 @@ module Basics::Lock {
 
     /// Lock something inside a shared object using a Key. Aborts if
     /// lock is not empty or if key doesn't match the lock.
-    public fun lock<T: store + key>(
+    public(script) fun lock<T: store + key>(
         obj: T,
         lock: &mut Lock<T>,
         key: &Key<T>,
@@ -108,7 +108,7 @@ module Basics::LockTest {
     }
 
     #[test]
-    fun test_lock() {
+    public(script) fun test_lock() {
         let user1 = @0x1;
         let user2 = @0x2;
