+++ sui_programmability/examples/defi/tests/EscrowTests.move
@@ -29,7 +29,7 @@ module DeFi::EscrowTests {
     }
 
     #[test]
-    public fun test_escrow_flow() {
+    public(script) fun test_escrow_flow() {
         // Both Alice and Bob send items to the third party
         let scenario = &mut send_to_escrow(ALICE_ADDRESS, BOB_ADDRESS);
         swap(scenario, &THIRD_PARTY_ADDRESS);
@@ -40,7 +40,7 @@ module DeFi::EscrowTests {
     }
 
     #[test]
-    public fun test_return_to_sender() {
+    public(script) fun test_return_to_sender() {
         // Both Alice and Bob send items to the third party
         let scenario = &mut send_to_escrow(ALICE_ADDRESS, BOB_ADDRESS);
 
@@ -63,7 +63,7 @@ module DeFi::EscrowTests {
 
     #[test]
     #[expected_failure(abort_code = 1)]
-    public fun test_swap_wrong_objects() {
+    public(script) fun test_swap_wrong_objects() {
         // Both Alice and Bob send items to the third party except that Alice wants to exchange
         // for a different object than Bob's
         let scenario = &mut send_to_escrow_with_overrides(ALICE_ADDRESS, BOB_ADDRESS, true, false);
@@ -72,14 +72,14 @@ module DeFi::EscrowTests {
 
     #[test]
     #[expected_failure(abort_code = 0)]
-    public fun test_swap_wrong_recipient() {
+    public(script) fun test_swap_wrong_recipient() {
         // Both Alice and Bob send items to the third party except that Alice put a different
         // recipient than Bob
         let scenario = &mut send_to_escrow_with_overrides(ALICE_ADDRESS, BOB_ADDRESS, false, true);
         swap(scenario, &THIRD_PARTY_ADDRESS);
     }
 
-    fun swap(scenario: &mut Scenario, third_party: &address) {
+    public(script) fun swap(scenario: &mut Scenario, third_party: &address) {
         TestScenario::next_tx(scenario, third_party);
         {
             let item_a = TestScenario::take_object<EscrowedObj<ItemA, ItemB>>(scenario);
