+++ sui_programmability/examples/nfts/tests/DiscountCouponTests.move
@@ -26,7 +26,7 @@ module NFTs::DiscountCouponTests {
     }
 
     #[test]
-    public fun test_mint_then_transfer() {
+    public(script) fun test_mint_then_transfer() {
         let scenario = &mut TestScenario::begin(&ISSUER_ADDRESS);
         {
             init(TestScenario::ctx(scenario));
