+++ sui_programmability/framework/sources/SUI.move
@@ -21,7 +21,7 @@ module Sui::SUI {
     }
 
     /// Transfer to a recipient
-    public fun transfer(c: Coin::Coin<SUI>, recipient: address, _ctx: &mut TxContext) {
+    public(script) fun transfer(c: Coin::Coin<SUI>, recipient: address, _ctx: &mut TxContext) {
         Coin::transfer(c, recipient)
     }
